requirements.txt
PyGithub
pydantic>=1.8.1,<2
